functions/image-processing/requirements.txt
flask
waitress
tox==3.*
Pillow
